Write to an SVG with a rect for each pixel, and show graph for debugging
The debug pixel graph and grid can be toggled with the constants defined in Main.cpp

There is an HTML file I added to the project directory that will display the SVG at many sizes. The workflow from now on should be to keep this open in your browser, then refresh every time you run our program.